refactor(autopilot): split service.go along its existing seams
service.go held the Run model, the Service, routing, source-transition
validation, bounded context assembly, summary rendering, and the run
invariants in 3062 lines — the largest file in the repository and the
biggest single cost to understanding this package.

Four concerns move out whole:

- run_context.go: bounded Action context assembly.
- run_summary.go: final summary rendering.
- source_transition.go: source refresh, amendment candidates, and the
  lineage and accepted-chapter rules, next to the tests already named
  source_transition_test.go.
- run_invariants.go: the state invariants and the destructive helpers
  they guard, with a note that they run on replay too, so a rule added
  there also judges every Run an earlier version wrote.

This is a pure move. Every non-blank line of the original file appears
exactly once across the five files and nothing else was added except the
four file headers; verified by comparing line multisets before and
after, and by the unchanged test suite.
